Work around unsquashed SerializeAfter instruction
In old times, SerializeAfter will trigger a squash, which is the most
``correct'' but conservative method to implement SerializeAfter syntax.
However, now GEM5 only stalls the successor instructions in decode
stages and rely on commit to squash. However, flushing SA
in commit stage depends on that the SA instruction is always
mispredicted. If ``correctly'' predicted (only for addr, not for
priv-level), no squash will be triggered.

So if we build an ultra-smart or lucky branch predictor, we will
encounter this bug. There are two ways to fix them:
- Force Rename to squash after SA instructions and stalls fetch until
the SA instruction is committed
- A Dirty & Cheating method: check all SA instruction in decode stage.
If they are predicted to be taken, mark them as not taken and stole
their predicted target.

We choose the second way in this patch