Refactor FTQ enquing and serving
Now even for a missing stream, it can only prediction limited lines.
This change avoids that fetching goes beyond stream prediction.
When fetching goes beyond stream prediction, we might have to
squash fetch from BP, which is annoying.

Because FTQ is aligned to cacheline and FSQ is not,
we must carefully maintain the last entry of FTQ.
There are 4 condtions:

A: Enque stream is taken, taken branch in current line
TakenPC is set to the start taken branch;
endPC is set to the end of taken branch.
For A, the target is completed right after enque.

B: Enque stream is taken, taken branch not in current line
C: Enque stream is (yet) not taken, endPC not in current line
TakenPC is set to 0; endPC is set to the start of next cache line.
For B&C, the target is completed right after enque.

D: Enque stream is (yet) not taken, endPC in current line
TakenPC is set to 0; endPC is set to the endPC of stream.
In this case (for D), the target is not completed after enque.
When we tryEnqFtq next time, we should fill up the last FTQ entry
with newly predicted info:
it might be reverted to taken or continue to be not taken.

(Note that fetching stops at ``endPC'' for a (yet) not taken stream,
because we don't know whether it will taken in next line.)