Fix get supplying FTQId and StreamId when fault
GEM5 fetch issues a pipelined cache acc at the end of the cycle.
When this cache acc is fault, it immediately build a nop.
When nop trys to obtain an FTQId and StreamId, it will trigger a
seg fault, because the FTQ entry is just used up and
supplying state set to nullptr.