cpu-pred: set default tables to alloc to 2
This is to avoid ping-pong on a useless main prediction entry:
Currently, we treate a replacement as successful allocation.
When two ping-pong outcome replacing each other in round-robin,
it will not allocate tables with longer history.

Set numTablesToAlloc=2 is only a workaround.
TODO: we should fix the ping-pong problem itself.