arch-riscv: Fix PTW congested by squashed txns
- The squash bandwidth of PTW is limited to 4 txns/cycle
The function to squash PTW is not called every cycle.
Instead, it was only called on PTW responsing (bug).
So, when there is excessive squashed txns, the number of
pending squashed txns will accumulate.
Translation contains a pointer pointing DynInstPtr,
which is ref-counted and limited below 1500.
This cause an assertion fault
- We fix this by scheduling the squashing function
every cycle when there is pending squash.