util: fix PGO scripts; fix IB-dependant code in addiw
- Old addiw implementation converts uint64 to int32,
which is implementation-defined and not standard defined.
To avoid compilers ``optimized'' based on UB, we cast it carefully.
- However, above fix does not prevent AOCC 4.1 from generating
wrong code that mistakenly simulates int32 instruction with int64 reg.
If I did not miss some UB, it's a compiler bug.
As a work around, we manually convert i32 X to u64 val:
If X > 0, u64 val = (u32) X; Else u64 val = (i64) X.

Change-Id: I2ae041d4da2180b3a70423eb99a7cdc02f8b39d1