cpu-o3: fix a memory leak problem
After the cache refills the data, it sends a `Bus_Clear` packet to notify the LSU.
This should only be sent by the Dcache, and neither L2 nor L3 should send this packet.

Calling `recvFunctionalCustomSignal` on the xbar does nothing,
causing the custom packet to be lost, leading to a memory leak.

Change-Id: Ie735660b66536908633daa6c25d76b7127add0b9