cpu-o3: pred: check histInfo match in ubtb replacement.
before this commit, the hisotry info of a S3 pred is directly trained into ubtb and its correctness will not be checked when trained with a new S3 pred, this results in a massive increase in intra flush caused by history info mismatch. Now, evertime a new S3pred gets trained into the ubtb, we check if the existing entry has the right amount of numNTCond, which is ubtb's way of storing the histInfo.

Change-Id: I88fb40b8efd7e7a5702b4dc915ce5a210121396e