cpu-o3: fix bug, add request validation in fetchCacheLine
This commit introduces a check in the fetchCacheLine function to determine if the ICacheHandler can accept new requests. If there are pending requests, it sets the appropriate stall reason and prevents further fetching until the requests are processed. Additionally, it enhances logging to provide better visibility into the FTQ state and cache request handling.

Change-Id: Ice826d8627404d4d9a142c4d8d1987c6668db845