cpu-o3: pred: make better use of mbtb's 2 taken.
mbtb's support of 2 taken was originally added for the validation of ubtb's 2nd pred. However, we later started using the mbtb's 2nd pred themselves, as a 2nd prediction when the validation fails. To push this idea one step further, we decided to make every tick a 2 taken tick (2taken ratio = 1). we do this by enqueueing mbtb's 2nd prediction even when there's no 2 taken hit, or the first prediction cause a override

Change-Id: I0e700732fc2df09bcaf4b3258778d37a52beb5dd