mem: Add pipe_latency parameter to BaseCache
This patch introduces a `pipe_latency` parameter to the BaseCache.

The primary purpose of this parameter is to delay the time when refill data is considered ready in the L1D cache.
During this latency window, load instructions can still forward data from the refill bus.
This mechanism allows the simulation to better align with the RTL refill timing.

Change-Id: Ib4ba8d561b511c4b473f5cdbcddd3294881a2340